Enhance ReferenceSolution to handle float32 inputs and update tests
- Modified the ReferenceSolution class to ensure that boundary coordinates are converted to float type when processed.
- Updated the evaluation method to accept float32 inputs for curved boundary queries.
- Added a new test case to validate the functionality of ReferenceSolution with float32 curved boundary queries.
- Updated binary reference solution file to reflect changes in the solution.

+++ src/deepflow/reference/solution.py
@@ -126,8 +126,18 @@ class ReferenceSolution:
                 continue
             bound_points = np.column_stack(
                 [
-                    np.asarray(bound_x.detach().cpu() if isinstance(bound_x, torch.Tensor) else bound_x),
-                    np.asarray(bound_y.detach().cpu() if isinstance(bound_y, torch.Tensor) else bound_y),
+                    np.asarray(
+                        bound_x.detach().cpu()
+                        if isinstance(bound_x, torch.Tensor)
+                        else bound_x,
+                        dtype=float,
+                    ),
+                    np.asarray(
+                        bound_y.detach().cpu()
+                        if isinstance(bound_y, torch.Tensor)
+                        else bound_y,
+                        dtype=float,
+                    ),
                 ]
             )
             if bound_points.size:
@@ -136,7 +146,12 @@ class ReferenceSolution:
         if not boundary_points:
             return inside
 
-        query_points = np.column_stack([np.asarray(x), np.asarray(y)])
+        query_points = np.column_stack(
+            [
+                np.asarray(x, dtype=float),
+                np.asarray(y, dtype=float),
+            ]
+        )
         scale = max(
             1.0,
             float(np.max(np.abs(query_points))) if query_points.size else 1.0,
